+++ frontend/config-assistant/src/components/gui-editor/ConfigTreeNodeResolver.ts
