+++ frontend/config-assistant/src/helpers/dataToString.ts
